refactor: utilsディレクトリを新設し、utils.jsを移動。理由: ユーティリティの管理を明確化し、拡張性を高めるため。
